Gydnc update for clojurescript failure
Attempting to add clojurescript hasn't gone well. It looks like it might
require a glibc based system -- but because we're alpine we get MUSL.
The fix is to compile from source, and went down a journey of trying to
use: https://github.com/anmonteiro/lumo and it's build instructions
didn't go so well. It seems like there is a JVM based clojurescript
compiler (interpreter?) that I'll take a swing with.