Enable fragment execution with Python
As I want these containers to support fragment invocation in an MCP
server. This would mean we could have like... 80? 90? languages
supported by an MCP server for running fragments. That's absurd. Also,
this provides a way for AI to **prove** stuff.

For example, you want a solution to the 8-queens? Prolog's code might be
most expressive, and maybe that's how AI should validate it's answer,
and present to a human.

Who knows, but this one works with AI now and it should extend
reasonably well.