A Fraglet Brainfuck
Chose brainfuck because... I have no idea how to create comments in it
so that the previous marker-based fraglet-entrypoint code injections
would work. This meant we'd have to update the entyrpoint to just
replace a file -- easy!

What's **really** fun about this, is because I have an MCP server that
can utilize fraglet compatible containers -- I asked AI to generate the
first 10 fibinacci numbers with brainfuck. It fell a part a little,
until it decided to use a python fraglet to generate a brainfuck string
and then used that brainfuck string to satisfy the requirements.

This is exactly what the world needed.