Transition to simpler fraglet pattern
We just install the fraglet-entrypoint, but only activate it whereever
it is implemented. Otherwise, we end up breaking some containers down
the dependency chain (for example, scala, arnoldc, and other containers
based on 100-java11)

Also, because we've ripped the entrypoint out -- we can nuke all the
prototype gocode that was left lingering around.